# control-plane/pkg/api/openapi.yaml
openapi: 3.1.0
info:
  title: Hanzo Agents Control Plane
  version: "1.0.0"
  summary: Per-org CRUD over the Brain and Bot CRDs.
  description: |
    The control plane is a thin REST to CRD facade. It authenticates
    callers via the IAM gateway, writes Brain and Bot custom resources
    to the management cluster, and reads their status back. The
    hanzoai/operator reconciles those CRs into Deployments, Services,
    and PersistentVolumeClaims inside the org's namespace.
servers:
  - url: https://agents.hanzo.ai
    description: Production
  - url: http://localhost:6430
    description: Local dev
components:
  securitySchemes:
    GatewayIdentity:
      description: |
        Identity headers injected by the IAM gateway (api.hanzo.ai)
        after JWT validation. Clients never set these directly; the
        gateway strips any client-supplied copies before forwarding.
      type: apiKey
      in: header
      name: X-Org-Id
  parameters:
    Org:
      name: org
      in: path
      required: true
      description: Org slug. Must match the caller's X-Org-Id, else 403.
      schema:
        type: string
        pattern: '^[a-z][-a-z0-9]{0,61}[a-z0-9]?$'
...
  schemas:
    ImageRef:
      type: string
      description: '"repo" or "repo:tag" reference (default tag "latest").'
      example: ghcr.io/hanzoai/brain:v1.2.3
    BrainPut:
      type: object
      additionalProperties: false
      properties:
        image:
          $ref: '#/components/schemas/ImageRef'
        replicas:
          type: integer
          minimum: 0
          maximum: 32
        storage:
          type: string
          example: 20Gi
        host:
          type: string
          example: brain.acme.hanzo.ai
    BrainView:
      type: object
      required: [name, namespace, spec, status]
      properties:
        name: { type: string }
        namespace: { type: string }
        spec:
          type: object
          additionalProperties: true
        status:
          type: object
          additionalProperties: true
...
    Error:
      type: string
  responses:
    Unauthorized:
      description: Missing or invalid identity headers.
      content:
        text/plain:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Path {org} did not match caller's X-Org-Id.
      content:
        text/plain:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource does not exist in the caller's namespace.
      content:
        text/plain:
          schema: { $ref: '#/components/schemas/Error' }
security:
  - GatewayIdentity: []
paths:
...
  /v1/orgs/{org}/brain:
    parameters:
      - $ref: '#/components/parameters/Org'
    get:
      operationId: getBrain
      summary: Read the org's Brain CR and status.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BrainView' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    put:
      operationId: putBrain
      summary: Create or update the org's Brain CR.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BrainPut' }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BrainView' }
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BrainView' }
        "400":
          description: Invalid body
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
    delete:
      operationId: deleteBrain
      summary: Tombstone the org's Brain CR.
      responses:
        "204": { description: Deleted }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
...